pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = ""
authors = ["jensjvh <43621451+jensjvh@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^2.2.0"
python-dotenv = "^1.0.1"
argon2-cffi = "^23.1.0"


[tool.poetry.group.dev.dependencies]
coverage = "^6.5.0"
pytest = "^8.3.3"
pytest-dotenv = "^0.5.2"
autopep8 = "^2.3.1"
pylint = "^2.15.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
